Add the rest of the API handlers
Most of these types are basic, with only one primary key, so the
generic operations work as expected. If you try to post without the
primary key set it auto generates, otherwise it attempts to use the
supplied key, not sure if I want to disable this. Additionally in order
to delete an item, the primary key has to be set, this is the expected
behavior, all without writing SQL. Need to test tables with
multi-column primary keys.